Experiment with error handling
Before each row would get a single status enum and if there was a
parsing error for any field, we'd bail on the entire row.
Now we try (naively) to recover from errors,
each row could have multiple flags set and every time
an error happened or a value was missing we set a flag in
`column_indicators`

This also means we allocate more `column_indicators` -- maybe we want
`fail-fast` option, where we won't try to recover from errors.