Make number of results match the number of tasks
This makes it easier to work with the results as it makes sure that none
of the result buffers could be overwritten before
the current chunk of bytes is done.

Ultimately we'll make the number of results configurable using channels,
but currently waiting on channels allocates and we don't want that.

We also fixed counting rows and displaying errors in debug context.